chore(ci): make lighthouse work in PR from forks
GitHub pull_request workflows are run within the context of the forked
repo and they can't comment on they own PR.

Using pull_request_target would be a security hole since we need to run
PR-provided arbitrary code to build the site.

The only viable option is `workflow_run` that detect when the
Lighthouse job has finished, as it is run in the context of the parent
repo.

That's extremely convoluted and not ideal because it doesn't even
provide full protection (see below), but it's a GitHub limitation
that needs to be fixed on their side with something like a
new 'comment-write' permission on ephemeral tokens.

There's still a way for a malicious PR to be annoying by
commenting on other PRs but it's not a security issue

Fix #530.