refactor: enable prettier and fix the styling
`prettier` needs to be configured in eslint with `"plugin:prettier/recommended"`,
as documented at
https://github.com/prettier/eslint-plugin-prettier#recommended-configuration

The previous config didn't really work and acted as if prettier was
disabled.

Also use `classnames` to avoid string concatenation. The linter
doesn't catch all the occurrences, so there may be places where
concatenation for classes is still there.

Fix #742
Fix #630
Fix #848