fix: refactor all stacking issues (z-index)
Remove the `z-index`es and rely on default stacking contexts
provided by Bootstrap Italia.

We should probably never use z-index, and if we do for special
cases, we should always use Bootstrap Italia variables.

The special case here is the sidebar - which in itself is a strange
component - and maybe we can remove the special casing for it as well
by adding the `.modal` class in JS when on mobile viewports, maintaining
the same semantics (in practice the sidebar turns into a modal on small
viewports).

In case there is something wrong in BI stacking in the future, we can fix
it there (https://github.com/italia/bootstrap-italia/issues/1009)

Fix #1038.